requirements.txt
flask==1.1.1
pymysql==0.9.3
sqlalchemy==1.3.2
redis==3.2.1
pyjwt==1.7.1
flask-cors==3.0.7
gunicorn==19.9.0
gevent==1.4.0
flask-sqlalchemy==2.3.2
flask-migrate==2.1.1
flask-script==2.0.6
flask-redis==0.3.0
flask-apscheduler==1.11.0
celery==4.4.0rc4
flask-celery-helper==1.1.0
